Bug fix for Android Q
It wasn't possible to use the CAB since the click and long
click listener weren't called. As new strategy now it's in
charge of recycler view dispatch events to the view holder
since the recycler view has alredy a touch listener to
dispatch item click and item long click to library users.